Re-org unit files + add spelunker/spirit guard
