+++ src/cardDb/units/mages/crystal.ts
@@ -1,0 +1,74 @@
+import { makeUnitCard as makeCard } from '@/factories/cards';
+import { UnitCard } from '@/types/cards';
+import { EffectType, TargetTypes } from '@/types/effects';
+import { Resource } from '@/types/resources';
+
+const THIRD_YEAR_STUDENTS: UnitCard = makeCard({
+    name: 'Third Year Students',
+    imgSrc: 'https://images.pexels.com/photos/8391400/pexels-photo-8391400.jpeg',
+    cost: {
+        [Resource.CRYSTAL]: 1,
+    },
+    description: '',
+    enterEffects: [],
+    totalHp: 4,
+    attack: 0,
+    numAttacks: 1,
+    isRanged: true,
+    isMagical: true,
+    isSoldier: false,
+    passiveEffects: [],
+});
+
+const MAGICIANS_APPRENTICE: UnitCard = makeCard({
+    name: "Magician's Apprentice",
+    imgSrc: '/images/units/magicians-apprentice.avif', // https://images.unsplash.com/photo-1615962122149-ef6987cf19f4'
+    cost: {
+        [Resource.CRYSTAL]: 1,
+        [Resource.GENERIC]: 1,
+    },
+    description: '',
+    enterEffects: [
+        {
+            type: EffectType.DRAW,
+            strength: 1,
+        },
+    ],
+    totalHp: 1,
+    attack: 1,
+    numAttacks: 1,
+    isRanged: true,
+    isMagical: true,
+    isSoldier: false,
+    passiveEffects: [],
+});
+
+const FAE_DRAGON: UnitCard = makeCard({
+    name: 'Fae Dragon',
+    imgSrc: 'https://cdn.pixabay.com/photo/2017/09/13/21/29/fantasy-2747066_1280.jpg', // https://images.unsplash.com/photo-1615962122149-ef6987cf19f4'
+    cost: {
+        [Resource.CRYSTAL]: 1,
+        [Resource.GENERIC]: 2,
+    },
+    description: '',
+    enterEffects: [
+        {
+            type: EffectType.DISCARD_HAND,
+            target: TargetTypes.SELF_PLAYER,
+            strength: 3,
+        },
+    ],
+    totalHp: 3,
+    attack: 3,
+    numAttacks: 1,
+    isRanged: true,
+    isMagical: true,
+    isSoldier: false,
+    passiveEffects: [],
+});
+
+export const CRYSTAL_MAGES = {
+    THIRD_YEAR_STUDENTS,
+    MAGICIANS_APPRENTICE,
+    FAE_DRAGON,
+};
