+++ src/cardDb/units/mages/index.ts
@@ -1,0 +1,11 @@
+import { CRYSTAL_MAGES } from './crystal';
+import { FIRE_MAGES } from './fire';
+import { WATER_MAGES } from './water';
+import { WIND_MAGES } from './wind';
+
+export const MAGES = {
+    ...CRYSTAL_MAGES,
+    ...FIRE_MAGES,
+    ...WATER_MAGES,
+    ...WIND_MAGES,
+};
