+++ src/cardDb/units/tribes/saharans.ts
@@ -1,0 +1,67 @@
+import { makeUnitCard as makeCard } from '@/factories/cards';
+import { UnitCard } from '@/types/cards';
+import { EffectType, TargetTypes } from '@/types/effects';
+import { Resource } from '@/types/resources';
+
+const FORTUNE_PREDICTOR: UnitCard = makeCard({
+    name: 'Fortune Predictor',
+    imgSrc: '/images/units/fortune-predictor.avif', // https://images.unsplash.com/photo-1508310621848-9b567e9621bf
+    cost: {
+        [Resource.CRYSTAL]: 1,
+        [Resource.IRON]: 1,
+    },
+    description: '',
+    enterEffects: [
+        {
+            type: EffectType.DISCARD_HAND,
+            strength: 1,
+            target: TargetTypes.ALL_OPPONENTS,
+        },
+        {
+            type: EffectType.DEAL_DAMAGE,
+            strength: 1,
+            target: TargetTypes.ALL_OPPONENTS,
+        },
+    ],
+    totalHp: 1,
+    attack: 1,
+    numAttacks: 1,
+    isRanged: true,
+    isMagical: true,
+    isSoldier: false,
+    passiveEffects: [],
+});
+
+const CAPTAIN_OF_THE_GUARD: UnitCard = makeCard({
+    name: 'Captain of the Guard',
+    imgSrc: 'images/units/captain-of-the-guard.avif', // https://images.unsplash.com/photo-1554772279-37573acdd183
+    cost: {
+        [Resource.CRYSTAL]: 1,
+        [Resource.IRON]: 1,
+        [Resource.GENERIC]: 1,
+    },
+    description: '',
+    enterEffects: [
+        {
+            type: EffectType.DISCARD_HAND,
+            strength: 2,
+            target: TargetTypes.SELF_PLAYER,
+        },
+        {
+            type: EffectType.DRAW,
+            strength: 3,
+        },
+    ],
+    totalHp: 3,
+    attack: 2,
+    numAttacks: 1,
+    isRanged: false,
+    isMagical: false,
+    isSoldier: true,
+    passiveEffects: [],
+});
+
+export const SAHARANS = {
+    FORTUNE_PREDICTOR,
+    CAPTAIN_OF_THE_GUARD,
+};
